email functionality
+ configured send order details company email
+ moved inquiry form to contact page
+ added mobile message for admin panel
+ separated php code to its own include files
