+++ src/admin/admin-panel.php
@@ -1,70 +1,4 @@
-<?php
-session_start();
-
-if(!isset($_SESSION["adminid"])){
-
-  header("location: admin-login.php");
-
-}
-else
-{
-
-  include("classes/Database.class.php");
-  include("classes/Product.class.php");
-  include("classes/ProductCon.class.php");
-
-  $cat_name = "";
-
-  function deleteFile($prod_image_file_old){
-      
-    $fileDestination = '../../assets/uploads/'.$prod_image_file_old;
-    $realFileDestination = realpath($fileDestination);
-
-    if(is_writable($realFileDestination)){
-      if(file_exists($realFileDestination)){
-        unlink($realFileDestination);
-      }
-    }
-  }
-
-  if (isset($_GET['id'])) {
-    $prod_id = $_GET['id'];
-
-    $product = new ProductController($prod_id);
-    $result = $product->getRecord($prod_id);
-
-    $row = $result->fetch(PDO::FETCH_ASSOC);
-    $prod_image_file_old = $row["prod_image_file"];
-
-    deleteFile($prod_image_file_old);
-
-    $error_msg = $product->removeRecord();
-    
-    if (empty($error_msg)) {
-      echo '<script>alert("Deleted Successfully!")</script>';
-    }else{
-      echo '<script>alert('.$error_msg.')</script>';
-    }
-  }
-
-  $product = new ProductController();
-  if(isset($_POST["submit"]))
-  {
-    // gets filtered table
-    $cat_name = $_POST["cat_name"];
-    
-    if(empty($cat_name)){
-      $result = $product->getTable();
-    }else{
-      $result = $product->getCatTable($cat_name);
-    }
-  }
-  else{
-    // gets normal table
-    $result = $product->getTable();
-  }
-}
-?>
+<?php include("includes/admin-panel.inc.php");?>
 
 <!doctype html>
 <html lang="en">
@@ -144,15 +78,28 @@ else
                     </tr>";
                 }
               ?>
-              
+
             </tbody>
           </table>
         </div>
-        
       </div>
-    </div>
+    </div>       
   </div>
 </section>
+
+<div class="container mobile-msg-wrapper text-center">
+  <div class="row d-flex mobile-msg justify-content-center mt-5 p-5">
+    <div class="col">
+      <h2>The Admin Panel is best experienced on a larger screen.</h2>
+      <p class="mt-5" style="font-size: 20px;">Please use a desktop or laptop to access this page. Thank you!</p>
+    </div>
+  </div>
+  <div class="row d-flex justify-content-center mt-5">
+    <div class="rc-admin-links">
+      <a href="admin-logout.php"><i class="bi bi-box-arrow-left btn-lg"></i>Logout</a>
+    </div>
+  </div>
+</div>
 
 </body>
