+++ src/admin/includes/admin-orders.inc.php
@@ -1,0 +1,15 @@
+<?php
+session_start();
+
+if(!isset($_SESSION["adminid"])){
+  header("location: admin-login.php");
+}
+else
+{
+  include("classes/Database.class.php");
+  include("classes/Order.class.php");
+  include("classes/OrderCon.class.php");
+
+  $order = new OrderController();
+  $result = $order->getTable();
+}
