+++ src/main/about.php
@@ -58,7 +58,4 @@
 <!-- Footer -->
 <?php include('footer.php') ?>
 
-<!-- Inquiry Button -->
-<?php include('inquiry.php') ?>
-
 </html>
