+++ src/main/checkout.php
@@ -109,9 +109,6 @@
 <!-- Footer -->
 <?php include('footer.php') ?>
 
-<!-- Inquiry Button -->
-<?php include('inquiry.php') ?>
-
 <!-- JS script -->
 <script src="checkout.js"></script>
