+++ src/main/contact.php
@@ -30,6 +30,14 @@
                     8 AM - 7 AM
                 </p>
             </article>
+            <article>
+                <h2>Inquire</h2>
+                <div>
+                    <a data-toggle="modal" data-target="#inquirychat" class="inquiry-btn">
+                    <img src="../../assets/speech-bubble.png">
+                    </a>
+                </div>
+            </article>
         </div>
         <div class="map-item">
             <h2>Find us!</h2>
@@ -45,7 +53,34 @@
 <?php include('footer.php') ?>
 
 <!-- Inquiry Button -->
-<?php include('inquiry.php') ?>
+<div class="modal fade" id="inquirychat" tabindex="-1" role="dialog" aria-labelledby="bannertopModalCenterTitle" aria-hidden="true">
+    <div class="modal-dialog modal-dialog-centered" role="document">
+        <div class="modal-content">
+            <div class="inquiry-modal">
+              <h2 class="mb-4"> Have a question?</h2>
+              <form class="inq-form" method="post" action="../../inquiry-send.php">
+                <br>
+                <div>
+                  <label for="name">Name:</label>
+                  <input type="text" id="name" name="name" placeholder="Enter your name" required>
+                </div>
+                <br>
+                <div>
+                  <label for="email">Email:</label>
+                  <input type="email" id="email" name="email" placeholder="Enter your email" required>
+                </div>
+                <br>
+                <div class="mb-4">
+                  <label for="subject">Subject:</label>
+                  <input type="text" id="subject" name="subject" placeholder="Enter message subject" required>
+                </div>
+                <textarea id="message" name="message" rows="5" cols="50" maxlength="1000" placeholder="Enter your message (max 1000 characters)" required></textarea>
+                <button type="submit" name="send" class="btn">Submit</button>
+              </form>
+            </div>
+        </div>
+    </div>
+</div>
 
 <map name="social-big">
     <area shape="rect" coords="0,3,40,43" href="https://web.facebook.com/RedVelvetKH">
