+++ src/main/includes/menu-cat.inc.php
@@ -1,0 +1,8 @@
+<?php
+
+include("../admin/classes/Database.class.php");
+include("../admin/classes/Product.class.php");
+include("../admin/classes/ProductCon.class.php"); 
+
+$product = new ProductController();
+$result = $product->getCatTable($current_cat);
