+++ src/main/menu-pastries.php
@@ -1,13 +1,4 @@
-<?php
-
-include("../admin/classes/Database.class.php");
-include("../admin/classes/Product.class.php");
-include("../admin/classes/ProductCon.class.php"); 
-
-$product = new ProductController();
-$result = $product->getCatTable("Pastries");
-
-?>
+<?php $current_cat = "Pastries"; include("includes/menu-cat.inc.php");?>
 
 <!doctype html>
 <html lang="en">
@@ -63,180 +54,10 @@ $result = $product->getCatTable("Pastries");
     </div>
 </section>
 
-<!-- <section class="menu-lay">
-    <div class="menu-card">
-        <div class="image-container">
-            <a data-toggle="modal" data-target="#pastrycinnamonModalCenter">
-                <img class="card-img" src="../../assets/food/pastry-cinnamon.jpg">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h4>Bestseller!</h4>
-                        <h2>Classic Cinnamon Rolls with Glaze</h2>
-                        <h3>$6</h3>
-                        <h4>(6 pcs)</h4>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h4>Bestseller!</h4>
-            <h2>Classic Cinnamon Rolls with Glaze</h2>
-            <h3>$6</h3>
-            <h3 style="font-size: 20px">(6 pcs)</h3>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a data-toggle="modal" data-target="#pastryensaymadaModalCenter">
-                <img class="card-img" src="../../assets/food/pastry-ensaymada.jpg">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Soft Ensaymada with Cheese</h2>
-                        <h3>$14</h3>
-                        <h4>(6 pcs)</h4>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Soft Ensaymada with Cheese</h2>
-            <h3>$14</h3>
-            <h3 style="font-size: 20px">(6 pcs)</h3>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a data-toggle="modal" data-target="#pastryyemaModalCenter">
-                <img class="card-img" src="../../assets/food/pastry-yemadonut.jpg">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Yema-Filled Donuts</h2>
-                        <h3>$14</h3>
-                        <h4>(8 pcs)</h4>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Yema-Filled Donnuts</h2>
-            <h3>$14</h3>
-            <h3 style="font-size: 20px">(8 pcs)</h3>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a data-toggle="modal" data-target="#pastryoldonutModalCenter">
-                <img class="card-img" src="../../assets/food/pastry-olddonut.jpg">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h4>Bestseller!</h4>
-                        <h2>Old-Fashioned Cheese Donuts</h2>
-                        <h3>$6</h3>
-                        <h4>(8 pcs)</h4>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h4>Bestseller!</h4>
-            <h2>Old-Fashioned Cheese Donuts</h2>
-            <h3>$6</h3>
-            <h3 style="font-size: 20px">(8 pcs)</h3>
-        </div>
-    </div>
-</section> -->
-
 </body>
 
 <!-- Footer -->
 <?php include('footer.php') ?>
-
-<!-- Inquiry Button -->
-<?php include('inquiry.php') ?>
-
-<!-- Menu Modal -->
-<!-- <div class="modal fade" id="pastrycinnamonModalCenter" tabindex="-1" role="dialog" aria-labelledby="pastrycinnamonModalCenterTitle" aria-hidden="true">
-    <div class="modal-dialog mw-100 w-50 modal-dialog-centered" role="document">
-        <div class="modal-content d-flex flex-row" id="menu-modal">
-            <img src="../../assets/food/pastry-cinnamon.jpg">
-            <div class="itm-modal-right">
-                <div class="itm-desc">
-                    <h2>Classic Cinnamon Rolls with Glaze</h2>
-                    <h5 style="font-weight:bold" class="mb-4">(6 pcs)</h5>
-                    <h4>Bestseller!</h4>
-                    <p>
-                        Lorem ipsum dolor sit amet, consectetur adipiscing elit.  
-                    </p> 
-                </div>
-                <div class="price-cart">
-                    <h3>$6</h3>
-                    <a class="btn" href="#" role="button">+ CART</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-<div class="modal fade" id="pastryensaymadaModalCenter" tabindex="-1" role="dialog" aria-labelledby="pastryensaymadaModalCenterTitle" aria-hidden="true">
-    <div class="modal-dialog mw-100 w-50 modal-dialog-centered" role="document">
-        <div class="modal-content d-flex flex-row" id="menu-modal">
-            <img src="../../assets/food/pastry-ensaymada.jpg">
-            <div class="itm-modal-right">
-                <div class="itm-desc">
-                    <h2>Soft Ensaymada with Cheese</h2>
-                    <h5 style="font-weight:bold" class="mb-4">(6 pcs)</h5>
-                    <p>
-                        Lorem ipsum dolor sit amet, consectetur adipiscing elit.  
-                    </p> 
-                </div>
-                <div class="price-cart">
-                    <h3>$14</h3>
-                    <a class="btn" href="#" role="button">+ CART</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-<div class="modal fade" id="pastryyemaModalCenter" tabindex="-1" role="dialog" aria-labelledby="pastryyemaModalCenterTitle" aria-hidden="true">
-    <div class="modal-dialog mw-100 w-50 modal-dialog-centered" role="document">
-        <div class="modal-content d-flex flex-row" id="menu-modal">
-            <img src="../../assets/food/pastry-yemadonut.jpg">
-            <div class="itm-modal-right">
-                <div class="itm-desc">
-                    <h2>Yema-Filled Donuts</h2>
-                    <h5 style="font-weight:bold" class="mb-4">(8 pcs)</h5>
-                    <p>
-                        Lorem ipsum dolor sit amet, consectetur adipiscing elit.  
-                    </p> 
-                </div>
-                <div class="price-cart">
-                    <h3>$14</h3>
-                    <a class="btn" href="#" role="button">+ CART</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-<div class="modal fade" id="pastryoldonutModalCenter" tabindex="-1" role="dialog" aria-labelledby="pastryoldonutModalCenterTitle" aria-hidden="true">
-    <div class="modal-dialog mw-100 w-50 modal-dialog-centered" role="document">
-        <div class="modal-content d-flex flex-row" id="menu-modal">
-            <img src="../../assets/food/pastry-olddonut.jpg">
-            <div class="itm-modal-right">
-                <div class="itm-desc">
-                    <h2>Old-Fashioned Cheese Donuts</h2>
-                    <h5 style="font-weight:bold" class="mb-4">(8 pcs)</h5>
-                    <h4>Bestseller!</h4>
-                    <p>
-                        Lorem ipsum dolor sit amet, consectetur adipiscing elit.  
-                    </p> 
-                </div>
-                <div class="price-cart">
-                    <h3>$6</h3>
-                    <a class="btn" href="#" role="button">+ CART</a>
-                </div>
-            </div>
-        </div>
-    </div>
-</div> -->
 
 </html>
