+++ src/main/menu.php
@@ -1,13 +1,4 @@
-<?php
-
-include("../admin/classes/Database.class.php");
-include("../admin/classes/Product.class.php");
-include("../admin/classes/ProductCon.class.php"); 
-
-$product = new ProductController();
-$result = $product->getTable();
-
-?>
+<?php include("includes/menu.inc.php")?>
 
 <!doctype html>
 <html lang="en">
@@ -63,123 +54,10 @@ $result = $product->getTable();
     </div>
 </section>
 
-
-<!-- <section class="menu-lay">
-    <div class="menu-card">
-        <div class="image-container">
-            <a href="menu-sigcakes.php">
-                <img class="card-img" src="../../assets/food/menu-sigcake.avif">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Signature Cakes</h2>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Signature Cakes</h2>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a href="menu-cakedel.php">
-                <img class="card-img" src="../../assets/food/menu-delcake.avif">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Cake Delights</h2>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Cake Delights</h2>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a href="menu-cheese.php">
-                <img class="card-img" src="../../assets/food/menu-cheese.avif">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Cheesecakes</h2>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Cheesecakes</h2></a>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a href="menu-pastries.php">
-                <img class="card-img" src="../../assets/food/menu-pastries.avif">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Pastries</h2>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Pastries</h2>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a href="menu-cupcakes.php">
-                <img class="card-img" src="../../assets/food/menu-cupcakes.avif">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Cupcakes</h2>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Cupcakes</h2>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a href="menu-cookies.php">
-                <img class="card-img" src="../../assets/food/menu-cookies.avif">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Cookies</h2>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Cookies</h2>
-        </div>
-    </div>
-    <div class="menu-card">
-        <div class="image-container">
-            <a href="menu-bars.php">
-                <img class="card-img" src="../../assets/food/menu-bars.avif">
-                <div class="overlay overlay--blur">
-                    <div class="overlay-content">
-                        <h2>Bars</h2>
-                    </div>
-                </div>
-            </a>
-        </div>
-        <div class="card-desc">
-            <h2>Bars</h2>
-        </div>
-    </div>
-</section> -->
-
 </body>
 
 <!-- Footer -->
 <?php include('footer.php') ?>
-
-<!-- Inquiry Button -->
-<?php include('inquiry.php') ?>
-
 
 </html>
