+++ src/main/order.php
@@ -71,9 +71,6 @@
 <!-- Footer -->
 <?php include('footer.php') ?>
 
-<!-- Inquiry Button -->
-<?php include('inquiry.php') ?>
-
 <!-- JS script -->
 <script src="cart.js"></script>
